internal: split up files stage into multiple files
Refactor the files stage to have the three main operations (passwd,
filesystem nodes, units) split into three new files.
